Replaced stategies with hub.options
- hub.LAST was the default anyway
- hub.CONCAT is replaced by hub.options({ allResults : true })
- this.options refers to the options object as is was passed to emit

Options are of type hub.Options and can be checked with instanceof. As a
consequence, the argument preceeding the callback can be a function that
will be passed on to the listeners.

Options properties are not directly exposed. Instead, an accessor
function is exposed which returns the value. This is to prevent users
from thinking that it's valid to change options in a listener.